Set send by email checkbox name per deployment environment
On the email settings page the "Send email by" checkbox in both the Test
and Live forms had the same label and id. This resulted in the user
being able to click on the label of a checkbox in a different
environment and it would affect the checkbox of the other environments
form.

Change the labels to be deployment environment specific so there is no
longer any clash.

In order to make this work once it is posted we extract the environment
specific parameter and set its value to the parameter expected by the
EmailSettings model which is `send_by_email`; 0 when unticked and 1 when
ticked.

This does mean that the Rails console output will complain about
unpermitted params for `send_by_email_dev` and `send_by_email_production`.