Generate service secret and token
The legacy publisher creates a service secret and service token for each
service which it then adds to the secrets.yaml during publishing.

We need to do the same here as the service token is used during file
upload to create the `encrypted_user_id_and_token` which the filestore
uses as the cipher key to encrypt the file before saving it to the S3
bucket.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>