Use app specific secrets
Further to the testable editor work, we also need to create a specific
secrets file for the testable branches as the editor_host is used for
the Auth0 implementation.

Without this app specific secrets the editor_host is overwritten each
time on deployment so the fb-editor-test ends authenticating into one of
the branch apps instead.