Update acceptance tests for changes in content component functionality
The content components have had a change in their functionality. They
are now activated when a user clicks on an `output` element which makes
an `input` element active. This `input` element is what actually
captures the user input.

In order for Capybara to be able to interact with content components as
close as possible to how user would we need to actively click on the
`ouput` elements before actually making changes to the `input` elements.