Set app specific config maps
The PR https://github.com/ministryofjustice/fb-editor/pull/481 was in
fact making the wrong thing app specific.

editor_host is indeed injected using the 'secrets' but it is actually
set in the `configmaps` not in the `secrets`. The secrets can be shared
across any instance of the editor as they are always the same.

The `configmaps` on the other hand _do_ need to be app specific :(