Parallelise lint, security and test runs in deployment pipeline
This changes the circleci config to use their base images and their ruby
and node orbs. Doing this enables dependencies to be cached between
steps. It also means the containers are faster to start up.

We therefore do not need to use docker-compose at all for running the
unit tests in the pipeline. This reduces the unit tests run from 7.5
minutes down to 1.5 minutes.

Bear in mind that the 1.5 minute assumes no changes to the Gemfile or
package.json. If there are changes to either then CircleCI will need to
reinstall those dependencies