Create session datastore adapter
This is mainly in response to the bug around optional checkbox answers.
When a user answers an optional checkbox and then returns to untick all
the answers this was not being reflected. This was fixed in the runner
however the editor's preview function uses a different method of saving
data.

This PR takes the editor closer to the runners approach of handling
data. the `AnswerParams` and `UserDataParams` are in fact exactly the
same in both apps. These could be moved to a single model in the
MetadataPresenter to be shared across both the editor and the runner.

Previewing an optional checkbox page should now save, and update user
data, as expected