Create flow objects for older services
Older services made use of the pages array their flow as since there was
no conditional logic the editor and runner would just traverse the array
in order.

The runner is backwards compatible and will make use of either flow
objects or an array of page objects.

This adds the functionality that creates the flow objects from the pages
array for services that do not already use the flow objects. This occurs
when the user clicks on the editor link for any of their given services.