Add flow object when adding new page to service
When a new page is added to a Service it currently either adds a page to
the end of the flow or inserts a page in the middle of the flow.

Add the required flow page objects to the service flow at the same time.
Since we are not doing branching at this stage when a page is added in
the middle of the flow it's default next should be the next page in the
flow. If added at the end then there should be no default next.

This requirement to always find an index for positionally adding a page
will go away once we have completed migration to using flow objects.