Use page UUIDs instead of IDs when updating and destroying
UUIDs are the identifier that we are predominantly using when
finding and interacting with pages and components. Using the ID is a
hold over from before UUIDs existed in the metadata.

With the soon to exist flow objects we use the UUID exclusively as the
object key as well as the reference to the previous and next objects.

When we come to update and destroy these flow objects we can no longer
use the page ID change the page updater and page destroyer to only
use UUIDs.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>