Delete flow objects when removing a page
When a user deletes a page from a service flow we also need to delete
the corresponding flow object for that page.

Since there is no conditional logic at this time the flow object that
links the object being deleted then needs to link the the following
object in the flow.