Always build service flow when in Test
Since we have testable editor apps which interact with the same database
as the regular fb-editor-test there is some cross over. A user in Test
may visit a testable editor which uses the flow objects but then go
back to the fb-editor-test and create or edit pages there.

Since the before_action only tests whether the flow object exists or not
any new pages would not be represented in the service flow. Therefore to
be on the safe side always rebuild the service flow when in Test.

Temporary until merged to main.