Base work for adding branching to the editor
We will have a Branch active model that will handle the conditionals
and expressions and will contain the validations as well.

And hard coded the params branch_id (which is a page uuid in the new
action and the flow_id in the edit)

The new and the edit action will require the previous flow uuid so
we can save the branch in the right place in the flow metadata.

In the form we will display the previous questions and all the pages

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>
Co-authored by: Brendan Butler <brendan.butler@digital.justice.gov.uk>