Return html for a new conditional
When the client side JS requests a new conditional we should
return sending the right name parameter.

So let's assume that on the adding branch page there are 3 conditionals
and the user clicks "Add another rule" / "Add another conditional" (I am
not sure what is the name of the button) then the JS will request
a new conditional sending the total amount of conditionals that exists
withing the page

/services/123/branches/456/conditionals/3 (3 conditionals)

and this will return a html that needs to be rendered within the page.