Set data attributes for supported form questions
Currently the only things that can be chosen on a conditional branch
object are questions radios and checkboxes. However we want to continue
showing all the questions that user may have created for their form.

When a user chooses a question that is not a radio or checkbox we should
show an error to the user. However there is no need for the frontend to
make a request to the backend in order to find out so instead we set
some data attributes on the option elements of the select list which
will help the frontend decide whether to show an error message or not.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>