Only return unique previous page objects
Currently the TraversedPages object in the presenter will just go and
find the next page object from the service flow regardless of how many
branch objects may or may not be pointing to it.

For the moment strip the duplicates out of the Branch object. We will
make changes to the TraversedPages object in the future to stop this
from happening.