Render Edit Branch page
This changes the expression answers template to allow for the rendering
of the branch edit page after the branch has been saved or the user is
directly visiting the edit page.

When a user creates a new branch it is necessary to show part of the
expression, specifically the list of previous questions that can be
chosen. However the answers and operators list should not be chosen.

When hitting the edit page we can just iterate over the answers as
normal but on the new page we have no answers to iterate over as the
user has yet to pick one.