Show previous questions on Edit Branch page
This is the basis for returning the question pages that are previous to
the point where a user is trying to add a branch.

Functionally the app will traverse all possible routes in a flow _after_
the point where the branch is being added, collate all the service pages
and then filter those pages out before returning them.

It will just return only pages that are of the single or multiple questions
type.

Currently this approach will return any question page which is
conceptually before the point in the flow the branch is being added.
This does include pages which would not technically be reachable should
a user of the form answer in such a way that they bypass the branch.

Co-authored-by Tomas Destefi <tomas.destefi@digital.justice.gov.uk>