Set service on each expression object in pages flow
The MetadataPresenter::Expression object has methods which get the
component and field for us but it relies on the Service object existing.

Currently the service is not passed into the intialiser of the
Expression so it is set at runtime whenever required.

Unless we change that we have no choice but to set the Service object on
the Expression each time we require access to the component and label.