Generate pages flow for service with branching objects
For services with branching we will need to build the vertical stacks of
pages.

If the previous object to the flow object being passed into a FlowStack
was a branch then the all the objects following that branch will be
grouped by their destination creating an array of arrays which will
functionally work like a 'stack' in the frontend.

It should be possible to expand this in the future to allow for branches
to be stacked as well as pages.

A better contextual summary of the work can be found in the PR message:
https://github.com/ministryofjustice/fb-editor/pull/666