Branching link work around
In order for these tests to work, we need to be able to access the branching page.
We currently do not have the add branching link in the live environment
so are not able to access the page this way.

This is a work around to access the branching page until the link becomes available.

Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>