Split acceptance test files by timings
Increase the parallelism to 4.

The acceptance tests will now be split across those 4 processes by
historical test run timing data which is uploaded and stored by CircleCi
after each run.

We do not bother doing this for the unit tests at this point as both the
ruby and javascript tests run in less than 60 seconds currently. If
there is seen to be a need to split them by time in the future then this
can be done using the acceptance tests split as the basis.

This reduces the accepatance tests run down from around 14-16 minutes to a
little over 4 minutes.