Add 'BRANCHING' feature flag
Add a feature flag called 'BRANCHING'. This will be used in the Test
environment, injected via the fb-editor-deploy repo.

This surfaces the link to 'Add branching' and also uncomments the
building of the new pages flow object.

We also changed the edit branch error summary acceptance tests to make
use of the link now available.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>
Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>