Order destinations list by flow
Previously we were ordering the pages and branches by when they were
added to the service. In fact these need to be ordered by the flow, much
in the same way as the PagesFlow object does when building the rich
object passed to the frontend.

Create an OrderedFlow object to be responsible for the correct ordering.
Use a DestinationsList module for creating the select list with the
correct attributes.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>