Update branching point destination list
We do not want the start page or confirmation page to appear in the
destination list.

Update the branch_spec to use the branching fixture as that has
branching point flow objects.

Update the destination_spec to have the correct ordering of flow
objects. The branches are no longer listed last.

There should be no duplicated list items.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>
Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>