Make sure that all items are generating UUIDs
New items added by the user are not generating the UUIDs, adding
the check makes the items generates the UUIDs.

Those are useful when we are selecting the expressions in the branching
If user chooses the option "I love Star Wars" (the UUID 1234-...) then
the user will be redirected to "Yes you shall pass!" page.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>