Add branch_uuid accessor
Currently, when a new branch is created it is not initialised with a branch_uuid.
Instead we use a previous_flow_uuid.

Only when a branch has been saved, is it assigned a branch_uuid.
However, we still refer to it as a previous_flow_uuid.
Conflating these two is confusing.

So we need to create a branch_uuid and reference it when we are dealing with an existing branch.