Subscribe to grid and route active support notifications
The Grid and Route models send an event notification should the number
of potential routes be exceeded of the number of service flow objects
be exceeded when building the grid. We will log these events for now and
see if something needs to be tweaked. If necessary we will send alerts to
Sentry.