Modify branch updater
Since we use a `branch_uuid` when the branch is being updated/edited and
`previous_flow_uuid` for when the branch is new/created, we will never have
a `previous_flow_uuid` in the BranchUpdater.

We also need to update the API calls to take both previous_flow_uuid and
branch_uuid.

Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>
Co-authored-by: Steven Burnell <steven.burnell@digital.justice.gov.uk>