Delete branch modal
This is the ground work for deleting a branch in the system.
The user will have to choose the destination route they want
to keep and the rest will turn into
unconnected/orphaned/detached pages.

This work doesn't include the actual deletion of a branch and will
be done in a future card.