Add a page after a branch
When adding a page after a branch we need to pass two params:

1. add page after
2. conditional uuid

The first is the branch and the second is the conditional route
where the new page will be inserted.

If we add a page after a branch and there is no conditional uuid
the code will look for the otherwise grid (the 'next' default).