Add grouping to branching destination lists
This PR adds a grouping for 'Unconnected pages' in the branching destination lists.

We decided to hand craft the erb templates rather than use the `govuk_select` gem
as the design calls for 'unconnected' options to be in an optgroup and 'connected'
options to not be in an optgroup. Although this design passes WCAG standards, mixing
options with and without a groupings inside a select html element is unusual.

Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>