Update Published services count by environment in admin dashboard
The way we were counting the actual published services per environment
needed to be updated to only look for the latest record for a distinct
service ID that had the status 'completed'.

Anything later with another status, such as 'unpublishing' or
'unpublished', means that it has been unpublished or is in the process
of being unpublished.