Temporarily remove branching acceptance tests
These tests will never pass when BRANCHING is set to disabled. We have
disabled it in the Test environment temporarily while we release a
version of the Editor that uses the new pages flow logic, but for
services without any branches.

Not use the `--exclude-pattern` because the the CircleCI config splits
the files by name to take advantage of parallelisation instead of just
running the entire acceptance tests suite in one command.