Add creation and destruction of uptime checks for each service
During publishing and unpublishing of a Service we need to create and
destroy Uptime Checks relating to that Service.

Currently we use Pingdom so there is a specific adapter relating to the
intereaction with their API.

When creating a new Pingdom check the url endpoint that we use is
/health. This is already used by Kubernetes to check the health of
services so we can just reuse this instead of creating another endpoint.