Only hide service URL when unpublished
Only checking for the `completed` status for a PublishService is not
enough as there is also other statuses that are used during the
publishing process.

Therefore hide the URL if the status is either `unpublishing` or
`unpublished` or if there is no PublishService associated with that
Service at all (i.e first publishing).

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>