Add legacy service name verification when creating a service
We need to ensure that users cannot generate the same URLs for a
form and as we allow people to define their form name this is a risk.

To prevent this from happening we currently check all form names on
MoJ Forms but we should also be checking forms on legacy Form Builder
too.

This piece of work is to build a simple validator to do this which
will enable a dynamic list to be updated by any member of the team
if and when a new form is added to legacy Form Builder.