Create Uptime Checks admin dashboard page
Instead of using the worker queues to automate the creation and deletion
of Uptime Checks we are going to use an admin dashboard page so a
developer can create and destroy them as required.

This page is only available in Live. Unpublishing is a manual process
anyway so we can add the destroying of an Uptime Check to the
unpublishing manual workflow.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>