Interpolate faraday version when stubbing in specs
This is a bit of a cheat but we do stub faraday requests in some of our
specs and the version is required. Since we have dependabot updating the
gem and assuming that a dev reviews it and merges the faraday update
then it's probably fine to trust the faraday version being interpolated
in the stub.