Add no branching editor deployment and acceptance tests
This adds the deployment steps necessary to create another editor with
container specific configuration which is more inline with what is
running in our Live environment.

We set the `DISABLE_BRANCHING` environment variable in the deploy step.
This is passed into the fb-deploy scripts and a extra argument called
`disable_branching` is passed to the helm command.

In the config_map.yml we check for the existence of that argument. If it
exists then we set the `BRANCHING` environemnt variable to be
'disabled'.

If you open that file in an editor that does any type of yaml file
parsing you will see that it looks like an error. This is not the case.
Helm is able to parse the if/else logic correctly.