Fixed typo possibly fixing preview bug
The preview function in the page menu had stoped working, throwing an error 
"TypeError: this.prevewPage is not a function. (In 'this.prevewPage(e)', 'this.prevewPage' is undefined)
in component_page_menu.js

I think it was just a typo, where preview was spelled without the i, "prevew".
This fixes that and hopefully restore the preview functionality in page menus.