Add modal for deleting branch destination with no default next
When a page which is the destination of a branching condition has no
default next if a user was to delete it there is nothing to take its'
place. This isn't really the state we want a branching point to ever get
to.

Instead we will check if the branch destination being deleted has a
default next. If it does not we show the user a modal letting them know
they cannot delete the page until they update the branch to point to a
different page manually.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>