Check existence of current_user object
The header partial makes use of the helper methods to check whether or
not to show the admin link. If there is no current user then calling
`email` will raise an error.

Add guard clauses to stop that from happening. If there is no
current_user then just return and do not show the admin link.