Update metadata when moving a flow object
Update the metadata to represent where a user wants to move an object to
in the flow.

If there is a previous flow uuid present then that previous objects
default next needs to be updated to point to what the object being moved
used to point to.

target_uuid represents the page that an object should be moved after or
the branch uuid for which the object being moved needs to replace either
one of the conditional destinations or the branch default next.