Require authentication when in the test environment
Changes in the templates reflect the requirement of setting a
username and password for all forms published in Test.

You can only publish a form without credentials in the Live environment,
this has been reflected in the unit test.

The publish modal is triggered with the use of the radio buttons
using JavaScript.