Create Enabled Validations Presenter
The majority of validations use a single modal when a user is
interacting with the configuration. However for the text and textarea
components their validation modals are grouped into pairs.

There is a separation between the min and max config, however the same
modal is used to set the character (*_length) and the word (*_word)
validations. As a result of this it is not possible to have a direct
mapping of validation type when the API is called.

In terms of the list of menu items instead of the text and textarea
components having:

- max_length
- min_length
- max_word
- min_word

they will have:

- max_string_length
- min_string_length

The BaseComponentValidation model will have to be able to assign the correct
component validation object during the API request.