Allow string length validator for text and textarea components
Because the text and textarea validation modals configure two different
types of validations (character and word count) the validator that will
be requested via the API will not map directly to the name of the
validator that the Editor will eventually set on the metadata.

As a result of this we need to allow requests with max_string_length or
min_string_lenght as the validator type for a text or textarea component
to validate in the BaseComponentValidation.