Re-publish live radio buttons
When a form user publishes their form to live without authentication,
and then clicks the 'Publish to Live' button again, neither radio buttons
are selected.

This is because the existing_authentication? method checks whether:
- A service has previously been published
- There is an existing authentication

We want the 'Set a username and password' radio to be selected when:
- A user has not previously published a form
- A user has published a form with authentication

We want the 'Allow anyone with the link to view' radio to be selected when:
- A user has published a form without authentication