Add preview for autocomplete component
To enable preview in the editor we require an API call to get
the items related to a component. This will be different in the
Runner, whereby the items will be present in an environment
variable.
The items will also be available in the autocomplete template
once a user uploads their items, as it may aid as a way to confirm
the uploaded items for form users.