Inject autocomplete items into container during publishing
This adds the ability for the publisher to request the autocomplete
items for a given service and then set them onto the environment
variable AUTOCOMPLETE_ITEMS before adding that to the deployment
configuration in Kubernetes.

For times when there are no autocomplete items for a service we just set
"{}" as the value. The Runners initialiser will just create an empty
hash on start up in those instances.