Add FromAddress Presenter and handle AWS SES errors
The FromAddress presenter will be used to ensure the correct message
will be passed to the front end.

Messages will be determined based on the email status; ie: deafult,
pending or verified.

We present accurate error messages to the front end if:
- AWS does not have the email submitted on record
- We aer unable to reach AWS.

When we call `get_email_identity` on the AWS SES client we expect to
receive an EmailServiceNotFound error if an email submitted by the
user doesn't exist in AWS yet.
If an AWS call fails we expect to get the EmailServiceError.