Add cron job to sync FromAddress table with AWS
We have a table that holds the verification status of a user's email.
To verify an email, a user needs to click a verification link sent to
their email address - this could occur at any time. Therefore, we
need a cronjob that interogates AWS SES API to ensure our database
reflects what is in AWS SES.