Add the initial Publishing Page Presenter
Create a Publishing Page Presenter to encapsulate all the different
possible warnings that we need to show our users.

We now pass in an array of potential messages for the from address and
submission pages presenters. Depending on the 'state' at any given
moment the presenters will decide whether or not show a message.

Renamed the presenters better describe what they do. There are now:

- Publishing Page Presenter - orchestrates all presenters
- Submission Pages Presenter - only deals with the Check Answers page
and Confirmation page
- From Address presenter - only deals with the From Address and the
potential states it can be in
- Autocomplete Items Presenter - relates to any autocomplete components
on pages and any issues they may have