Update Publishing Page acceptance tests
We have heavily changed the behaviour of the Publishing page.
Therefore, we have decided to re-write the acceptance tests for this
page.

We now have a catch all warning based on whether the user has their
'Send by Email' enabled.

If the 'Send by Email' has been enabled, we show an overall
warning message, which is followed by more specific errors that
relate to the submission pages, from address and autocomplete
errors.

Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>