Cater for autoclosing Configure summary link
In the publishing page acceptance spec we go to the email settings page
in order to configure the service email output address. Clicking on
'Configure' opens the form, but then it closes it again.

Therefore when the test was unable to find the field it needed it
failed. The terrible solution we implemented will capture the Capybara
exception and try one more time.

This is not a great solution and does increase the test time
significantly.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>